feat(pwa) make sure prod build runs after ssg. Provide css from ssg run to prod SW
